requirements.txt
alembic==1.7.6
certifi==2021.10.8
charset-normalizer==2.0.11
click==8.0.3
Flask==2.0.3
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
joblib==1.1.0
Levenshtein==0.17.0
Mako==1.1.6
MarkupSafe==2.1.0
nltk==3.6.7
numpy==1.22.2
python-dotenv==0.19.2
rapidfuzz==1.9.1
regex==2022.1.18
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
SQLAlchemy==1.4.31
threadpoolctl==3.1.0
tika==1.24
tqdm==4.62.3
urllib3==1.26.8
Werkzeug==2.0.3
